docs: 문서 추가 feature/TODAK-242/refactoring-document
