+++ domain/auth/components/signupSteps/PolicySection.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { RootState } from "@/redux";
+import { RootState } from "@/domain/shared/redux";
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { SIGNUP_STEP } from "../../constants";
