+++ domain/auth/components/signupSteps/SignupStepper.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { RedirectSVG } from "@/domain/shared/components/svg";
-import { RootState } from "@/redux";
+import { RootState } from "@/domain/shared/redux";
 import React from "react";
 import { useSelector } from "react-redux";
 import { SIGNUP_STEP } from "../../constants";
