+++ domain/shared/redux/index.ts
